fix(messages-moved): destination uses per-message badge + context menu, fix order
Two follow-up fixes after staging review:

1. **Drawer message order was reverse-chronological.** Backend now sorts
   `selectedMessages` by `sequence` ascending before building the
   `messages` array on the tombstone payload. The drawer renders messages
   in the order they were originally sent, not the order the user
   happened to click checkboxes in (the validate test passes
   `[movedB, movedA]` deliberately to lock this in).

2. **Destination tombstone is no longer rendered inline.** The destination
   stream already shows the moved messages themselves, plus a per-message
   origin badge — duplicating the tombstone inline was visual noise on
   top of the messages the user is already looking at. Replaced with:

   - `MessagesMovedEvent` returns `null` when `streamId === payload.destinationStreamId`
   - New "Show move details" entry in the per-message context menu, only
     visible when the message has `movedFrom` provenance and the destination
     tombstone is hydrated in IDB
   - The same `MovedMessagesDrawer` (now extracted to its own file) opens
     from both the source-side tombstone summary AND the per-message menu

   To make the destination drawer route work, each relocated `message_created`
   payload now carries `movedFrom.moveTombstoneId` — the destination
   tombstone's `event.id`. Pre-generated up front so it can be stamped via
   the same UPDATE that moves the row, without a second round-trip.

Other touch-ups while here:
- `MovedMessagesDrawer` is now self-contained (takes the tombstone event
  + workspaceId, calls `useActors` internally). Both call sites mount it
  conditionally so the IDB subscription only runs when the drawer is open.
- New `useMovedTombstone(id)` hook in `apps/frontend/src/hooks/` looks the
  destination tombstone up by id from the IDB events cache. Components
  can't import `@/db` directly (INV-15), so the hook lives in the hooks
  layer.

Verified: typecheck clean across all 9 workspaces; 47 affected frontend
tests still pass.
