+++ apps/frontend/src/hooks/index.ts
@@ -79,6 +79,7 @@ export {
 export { useSearch } from "./use-search"
 
 export { useActors, type ActorLookup } from "./use-actors"
+export { useMovedTombstone } from "./use-moved-tombstone"
 
 export { useWorkspaceEmoji } from "./use-workspace-emoji"
